nixos/caddy: improve security about acme certs
Before this patch, the caddy user will be added to the acme group if
acme is used, which may give caddy more permission than needed and is
inconsistent with the documentation.

This patch improves this to only give caddy the needed permission by
suggesting setting the group of needed certs to caddy, which is what
other web servers, e.g. nginx, do.

It also suggests the old way in case needed.