elisp-packages: fix AOT native-comp for several packages
Currently, AOT native-comp for emacs packages is implemented[1] in
postInstall. These packages override postInstall and do not keep
old.postInstall, so they do not do AOT native-comp. This patch fixes
this by keeping the old.postInstall in their overrides.

Note that I do not fix emacspeak because I have not found an elegant
way to do so.

[1]: a7cb8e36ac17526ffc2255c9c48c94be675c4d67